requirements.txt
#tensor2tensor[tensorflow_gpu]
#torch == 1.11.0
#numpy == 1.21.5
#gym[atari, box2d, accept-rom-license] == 0.23.1
gym[atari, accept-rom-license] == 0.23.1
#box2d-py
plotly
ipykernel
pandas
ipywidgets
cloudpickle
scikit-image
#scikit-learn
keras-rl2
Pillow
h5py
tensorflow==2.10.1
tensorflow-gpu==2.10.1
tensorflow-intel==2.10.1
keras==2.10.1
#tensorflow_addons 